Alter movie helper to style textarea elemets
Materialize requires text areas to have the "materialize-textarea" class
or they will not receive any styling.

This adds that class if the type in the database is :text, it seems safe
to assume that text areas would be used for collecting data stored in
the text type, and that other types (varchar) would not use text areas.

This does break non text area fields (text) visually. So all text type
items from a movie object should either use text area, or not be
rendered.