Alter movie helper to style datetime elements
Materialize allows you to use any class for date pickers, but it needs
javascript to actually set it up. 'datepicker' is the class used in
commit 954b080a where the javascript for datepickers was added. So that
class will be added to any field that is the type datetime on the movies
model. Datetime fields should only use datepickers as they force correct
format and dissallow users from inputting dates in incorrect formats.

Additionally datepickers do not support html5 validations. So this will
completely overwrite the out variable with the datepicker class. As such
this line sohuld always come after only classes that should not be on
datepickers. currently that is only the validate class